ThesisRecordBuilder/src/requirements.txt
python-docx
pillow
fitz
PyMuPDF
pytesseract
opencv-python
deskew
lxml
tqdm